requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
openai
requests==2.31.0
python-dotenv==1.0.0
pydantic-settings==2.0.3

pydantic~=2.8.2
langchain~=0.2.14
pip~=24.2
attrs~=24.2.0
distro~=1.9.0
wheel~=0.41.2
setuptools~=68.2.0
packaging~=24.1
zipp~=3.20.0
typing_extensions~=4.12.2
build~=1.2.1
jinja2
faiss-cpu
streamlit

